Move files not in ui project back to root
