+++ src/Buffer.test.ts
@@ -7,7 +7,7 @@ import { assert, expect } from 'chai';
 import { ITerminal } from './Types';
 import { Buffer, DEFAULT_ATTR_DATA } from './Buffer';
 import { CircularList } from './common/CircularList';
-import { MockTerminal, TestTerminal } from './ui/TestUtils.test';
+import { MockTerminal, TestTerminal } from './TestUtils.test';
 import { BufferLine, CellData } from './BufferLine';
 
 const INIT_COLS = 80;
