+++ src/Clipboard.test.ts
