+++ src/Linkifier.ts
@@ -3,9 +3,9 @@
  * @license MIT
  */
 
-import { IMouseZoneManager } from './ui/Types';
+import { IMouseZoneManager } from './Types';
 import { ILinkifierEvent, ILinkMatcher, LinkMatcherHandler, ILinkMatcherOptions, ILinkifier, ITerminal, IBufferStringIteratorResult } from './Types';
-import { MouseZone } from './ui/MouseZoneManager';
+import { MouseZone } from './MouseZoneManager';
 import { getStringCellWidth } from './CharWidth';
 import { EventEmitter2, IEvent } from './common/EventEmitter2';
