+++ src/MouseHelper.test.ts
